chore: refactor navigation to DRY
